cmdline: move install subcommand to separate file
It's a large enough struct, and has enough support code and tests, to
maintain separately.
